refactor(pojo): 批量更新实体类并新增大量业务实体
【变更内容】
- 优化CrWarnRule实体类，新增类注释和索引说明
- 修复sysUserProfile类名大小写错误，补充时间字段和隐私级别类型
- 更新SysUser、SysLoginLog、PsyScale等多个现有实体类的字段类型
- 新增PsyOption、PsyAnswer、CsChatMessage等20+业务实体类
- 完善README.md文档，补充项目说明和使用指南

【影响范围】
所有依赖实体类的业务层、数据访问层代码

+++ mhp-pojo/src/main/java/com/mhp/entity/CrWarnEvent.java
@@ -4,10 +4,6 @@ import lombok.Data;
 import lombok.Builder;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
-
-import lombok.NoArgsConstructor;
-import lombok.AllArgsConstructor;
-
 import java.time.LocalDateTime;
 
 /**
@@ -22,7 +18,7 @@ public class CrWarnEvent {
     // 主键
     private Long eventId;
     // 学生ID
-    private Integer studentId;
+    private Long studentId;
     // 评估ID
     private Long assessmentId;
     // 事件级别：1=低 2=中 3=高
@@ -32,7 +28,7 @@ public class CrWarnEvent {
     // 触发来源（如EXAM）
     private String triggerSource;
     // 处理人ID
-    private Integer handlerId;
+    private Long handlerId;
     // 创建时间
     private LocalDateTime createTime;
 }
